Moves useUpdateProfileImage into correct folder
