+++ packages/react/src/profile/index.ts
@@ -9,7 +9,7 @@ export * from './useExploreProfiles';
 export * from './useSearchProfiles';
 export * from './useCollectedPublications';
 export * from './useMutualFollowers';
-export * from './useUpdateProfileImage';
+export * from '../transactions/useUpdateProfileImage';
 export * from './useProfilesOwnedBy';
 
 export type {
