+++ packages/react/src/transactions/adapters/ProfileImageCallGateway.ts
@@ -23,11 +23,8 @@ import { SupportedTransactionRequest } from '@lens-protocol/domain/use-cases/tra
 import { ChainType, failure, success } from '@lens-protocol/shared-kernel';
 import { v4 } from 'uuid';
 
-import {
-  AsyncRelayReceipt,
-  ITransactionFactory,
-} from '../../transactions/adapters/ITransactionFactory';
 import { UnsignedLensProtocolCall } from '../../wallet/adapters/ConcreteWallet';
+import { AsyncRelayReceipt, ITransactionFactory } from './ITransactionFactory';
 
 export class ProfileImageCallGateway implements IProfileImageCallGateway {
   constructor(
